Split out QtHost.h from HostTypes.h
