+++ CMakeLists.txt
@@ -733,6 +733,8 @@ elseif(USING_QT_UI)
       ${QT_UI_GEN}
       Qt/QtMain.cpp
       Qt/QtMain.h
+      Qt/QtHost.cpp
+      Qt/QtHost.h
       Qt/mainwindow.cpp
       Qt/mainwindow.h
       Qt/Debugger/ctrldisasmview.cpp
