+++ Qt/QtHost.cpp
@@ -1,0 +1,29 @@
+// Copyright (c) 2014- PPSSPP Project.
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, version 2.0 or later versions.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License 2.0 for more details.
+
+// A copy of the GPL 2.0 should have been included with the program.
+// If not, see http://www.gnu.org/licenses/
+
+// Official git repository and contact information can be found at
+// https://github.com/hrydgard/ppsspp and http://www.ppsspp.org/.
+
+#include <string>
+#include "Qt/QtHost.h"
+
+const char* QtHost::SymbolMapFilename(std::string currentFilename) {
+	std::string result = currentFilename;
+	size_t dot = result.rfind('.');
+	if (dot == result.npos)
+		return (result + ".map").c_str();
+
+	result.replace(dot, result.npos, ".map");
+	return result.c_str();
+}
